Convert `src/redux/requests` to TypeScript.
